Seperate Client and Server safe calls
