+++ src/main/kotlin/com/the9grounds/aeadditions/core/mixin/CraftingCubeModelMixin.java
@@ -2,7 +2,7 @@ package com.the9grounds.aeadditions.core.mixin;
 
 import appeng.block.crafting.AbstractCraftingUnitBlock;
 import appeng.client.render.crafting.CraftingCubeModel;
-import com.the9grounds.aeadditions.util.AE2;
+import com.the9grounds.aeadditions.util.AE2Client;
 import net.minecraft.client.renderer.block.model.ItemOverrides;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.client.resources.model.BakedModel;
@@ -42,11 +42,11 @@ public abstract class CraftingCubeModelMixin {
     private void bakeExtraCpus(IModelConfiguration owner, ModelBakery bakery,
                                Function<Material, TextureAtlasSprite> spriteGetter, ModelState modelTransform,
                                ItemOverrides overrides, ResourceLocation modelLocation, CallbackInfoReturnable<BakedModel> callbackInfoReturnable) {
-        AE2.bakeExtraCpus(type, LIGHT_BASE, RING_CORNER, RING_SIDE_HOR, RING_SIDE_VER, spriteGetter, callbackInfoReturnable);
+        AE2Client.bakeExtraCpus(type, LIGHT_BASE, RING_CORNER, RING_SIDE_HOR, RING_SIDE_VER, spriteGetter, callbackInfoReturnable);
     }
     
     @Inject(method = "getAdditionalTextures()Ljava/util/stream/Stream;", at = @At("RETURN"), cancellable = true, remap = false)
     private void returnExtraCpuTextures(CallbackInfoReturnable<Stream<Material>> callbackInfoReturnable) {
-        callbackInfoReturnable.setReturnValue(AE2.combineTextures(callbackInfoReturnable.getReturnValue()));
+        callbackInfoReturnable.setReturnValue(AE2Client.combineTextures(callbackInfoReturnable.getReturnValue()));
     }
 }
